pyproject.toml
[tool.poetry]
name = "fastapi-mongodb-boilerplate"
version = "1.0.0"
description = "Production-ready FastAPI boilerplate with MongoDB, Redis, Celery, Clean Architecture & DDD"
authors = ["keniiy <kehindekehinde894@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = ">=0.115.0,<0.116.0"
uvicorn = ">=0.32.0,<0.33.0"
gunicorn = ">=23.0.0,<24.0.0"
motor = ">=3.6.0,<4.0.0"
pymongo = ">=4.10.0,<5.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
pydantic-settings = ">=2.6.0,<3.0.0"
pyjwt = ">=2.9.0,<3.0.0"
slowapi = ">=0.1.9,<0.2.0"
argon2-cffi = ">=23.1.0,<24.0.0"
email-validator = ">=2.2.0,<3.0.0"
python-json-logger = ">=2.0.7,<3.0.0"
# Redis
redis = ">=5.0.0,<6.0.0"
# Celery
celery = ">=5.4.0,<6.0.0"
flower = ">=2.0.0,<3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.0,<9.0.0"
pytest-asyncio = ">=0.24.0,<0.25.0"
pytest-cov = ">=6.0.0,<7.0.0"
httpx = ">=0.28.0,<0.29.0"
# Code Quality
black = ">=24.0.0,<25.0.0"
isort = ">=5.13.0,<6.0.0"
ruff = ">=0.5.0,<0.6.0"
mypy = ">=1.10.0,<2.0.0"
pre-commit = ">=3.7.0,<4.0.0"
bandit = ">=1.7.0,<2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError"
]

# Black code formatter
[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# isort import sorter
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", "*/__pycache__/*"]

# Ruff linter
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "ARG001", # unused function argument (common in FastAPI routes)
    "ARG002", # unused method argument
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/*" = ["ARG", "S101"]  # allow assert in tests

# MyPy type checker
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "app.infrastructure.db.*",
    "motor.*",
    "pymongo.*",
]
ignore_missing_imports = true

